pyproject.toml
[project]
name = "lab01-part2-13775103"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "kagglehub[pandas-datasets]>=1.0.0",
    "matplotlib>=3.10.8",
    "pandas>=3.0.2",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=23.0.1",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
]
